Изменения: исправила background, теперь закрашивается весь фон; исправила стили в App использовала стейты вместо defaultArticleState; также исправила моменты Можно лучше
