+++ src/components/app/App.tsx
@@ -1,0 +1,26 @@
+import { CSSProperties, useState } from 'react';
+import { defaultArticleState } from 'src/constants/articleProps';
+import { ArticleParamsForm } from '../article-params-form';
+import { Article } from '../article';
+import styles from 'src/styles/index.module.scss';
+const App = () => {
+	const [pageStyle, setPageStyle] = useState(defaultArticleState);
+	return (
+		<main
+			className={styles.main}
+			style={
+				{
+					'--font-family': pageStyle.fontFamilyOption.value,
+					'--font-size': pageStyle.fontSizeOption.value,
+					'--font-color': pageStyle.fontColor.value,
+					'--container-width': pageStyle.contentWidth.value,
+					'--bg-color': pageStyle.backgroundColor.value,
+				} as CSSProperties
+			}>
+			<ArticleParamsForm setPageStyle={setPageStyle} />
+			<Article />
+		</main>
+	);
+};
+
+export default App;
